Docs : Nfc 관련 문서 구조화
Nfc 자바스크립트 파일 component, container, screen으로 분할하여 구현
